entry: Fake a zImage Header
A zImage is not a raw executable but has a header located at the offset 0x24,
which contains a magic, start and end addresses, and some endianness
information.

If this header is missing, bootloader might get confused, defeating the goal of
having babelfish running in any cases, and behaving like a real kernel.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>