head: Initialize the stack pointer
If we've just been booted in non-secure, we're running without any stack.
Make sure we have one.

To do so, create a new section in our binary so that we know no-one will step
on our foot, yet without providing any absolute address that might not apply to
our case.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>