babelfish: Allow the binary to be position independant
So far, the binary was generated with the assumption that it was going to be
run at the address 0x40008000.

This was convenient because quite simple to do, but wasn't really pratical,
since for a number of reasons (u-boot would not relocate our kernel,
Allwinner's bootloader would load us at a different address, the A80 seems to
have a different RAM base address, etc.) we might not be running from there.

Implement PIC in several steps:
  - Remove the text section address assignment
  - Discard the .data* sections, since any r/w data will generate a GOT based
    relocation, which won't work for us. This will also trigger a link
    error any time we will actually use one, which is good.
  - Remove the global variables that were used before, mostly the malloc_base
    variable, and create a new section for the heap instead, that will be used
    to hold our allocated memory. Since we can't use C global variables, the
    global heap_base variable will be set from the assembly code.
  - Convert fdt_strerror to a dumber but global-free code in order to
    accomodate with the above constraints
  - Retrieve the execution base address, and use that when dereferencing now
    relative pointers.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>