main: Apply Cortex-A7 init quirks when running on such a CPU
The Cortex-A7 were relying a bit more on the firmware to set up a proper
environment and a few registers.

For example, the SMP bit is needed to enable the caches, and CNTFRQ to set the
arch timers frequency.

Since the Allwinner bootloader was not doing such a job, do it ourselves.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>